Simplify CLI options and set default driver
This change removes the driver options and turns them into flags which
should make it easier for users to specify settings. It still maintains
the generic "nature" of how the config options work with multiple drivers,
so it would in theory be possible to add different drivers other than
the default "kubernetes" driver in the future.