Fix containerd socket mount on retry
Recent UX refactoring broke part of the logic for related to runtime retries
during default creation.  The result was the builder "worked" but didn't
actually have a containerd socket mount and was running with the worker
instead.  This also refines the basic integration test to run a Pod with the
image that was just built.