Add support for custom config mounting
This adds support to mount a custom ConfigMap containing files
into /etc/config which can then be used for tuning the buildkitd
configuration.  For example, if your registry uses self-signed or
non-standard certs, you can pass those in and configure buildkitd
to trust them.

This also fixes a subtle glitch where we should be passing in the root
path as a flag instead of config file setting, thus making it easier for
users to BYO config file and not run into problems.