Refactor BuildKit Image definition
Instead of having the image buried in the repo, lets move it up to the
version package to make it easier to find and update.  We may elect
to wrap the image to inject additional helpers in the future.

This also fixes the rootless image to use the latest stable upstream image